requirements.txt
psutil>=5.9.0
watchdog>=3.0.0
scikit-learn>=1.3.0
torch>=2.0.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.35.0
pyyaml>=6.0
